fix(rog-slash): drop repr(u8) so DeviceState keeps its (byyu) signature

+++ asusctl/src/slash_cli.rs
@@ -97,7 +97,7 @@ pub fn handle_slash_set(cmd: &SlashSetCommand) -> Result<(), Box<dyn std::error:
             proxy.set_interval(interval)?;
         }
         if let Some(slash_mode) = cmd.mode {
-            proxy.set_mode(slash_mode)?;
+            proxy.set_mode(slash_mode as u8)?;
         }
         if let Some(show) = cmd.show_on_boot {
             proxy.set_show_on_boot(show)?;
